# Optimization: asynchronous writeback