@sourcejedi 2015-08-12
Obviously there's a few lines of code here, but it's not intrusive.  I tried to keep it as simple as possible.

It's still slightly clever, but there's no point using python if you're not going to pass a few closures between threads :).